Update livestream_with_jetson-v0.18
No difference to main branche.
I removed a jpegparse, which caused stutter and freezes with the webcam. The problem was, that it had worked before. GStreamer is a fixed video/audio pipeline. The jpegparse plugin buffers one frame. GStreamer calculates for the complete path the runtime and delays and latency to the sinks. Maybe it has to do with this one frame in the jpegparse element.